Add draw line feature
This feature draws line from the FinalLine coordinates provided.
Feature is not yet complete. Also currently lines are drawn of
white colour, instead we have to increment the value of pixel by
one to count number of lines at a place. Also we have to draw lines
extending both sides, till now we are only extending on one side.